requirements.txt
asttokens==2.4.1
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
executing==2.0.1
fonttools==4.46.0
freetype-py==2.4.0
glfw==2.6.3
icecream==2.1.3
imgui==2.0.0
importlib-resources==6.1.1
kiwisolver==1.4.5
matplotlib==3.8.2
multipledispatch==1.0.0
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
Pygments==2.17.2
PyOpenGL==3.1.7
pyparsing==3.1.1
pyrr==0.10.3
python-dateutil==2.8.2
six==1.16.0
zipp==3.17.0
